+++ packages/terra-progress-bar/src/ProgressBar.jsx
@@ -27,12 +27,17 @@ const propTypes = {
    * Boolean check for gradient.
    */
   hasGradient: PropTypes.bool,
+  /**
+   * Value passed to aria-valuetext for accessibility
+   */
+  valueText: PropTypes.string,
 };
 
 const defaultProps = {
   heightSize: 'small',
   max: 100,
   hasGradient: false,
+  valueText: undefined,
 };
 
 const ProgressBar = ({
@@ -40,6 +45,7 @@ const ProgressBar = ({
   value,
   max,
   hasGradient,
+  valueText,
   ...customProps
 }) => {
   const classes = cx([
@@ -61,6 +67,7 @@ const ProgressBar = ({
       aria-valuemax={100}
       aria-valuemin={0}
       aria-valuenow={normalizedValue}
+      aria-valuetext={valueText}
       tabIndex="-1"
     />);
 };