+++ packages/terra-progress-bar/src/terra-dev-site/test/progress-bar/ProgressBarColor.test.jsx
@@ -3,13 +3,13 @@ import ProgressBar from '../../../ProgressBar';
 
 const ProgressBarColor = () => (
   <div>
-    <ProgressBar id="progressbarWithNamedColor" value={5} max={10} color="Orange" />
+    <ProgressBar id="progressbarWithNamedColor" valueText="5%" value={5} max={10} color="Orange" />
     <br />
     <br />
-    <ProgressBar id="progressbarWithHexColor" value={7.5} max={10} color="#8ccc62" />
+    <ProgressBar id="progressbarWithHexColor" valueText="7.5%" value={7.5} max={10} color="#8ccc62" />
     <br />
     <br />
-    <ProgressBar id="progressbarWithRGBColor" value={10} max={10} color="rgb(255, 0, 0)" />
+    <ProgressBar id="progressbarWithRGBColor" valueText="10%" value={10} max={10} color="rgb(255, 0, 0)" />
   </div>
 );