+++ packages/terra-progress-bar/src/terra-dev-site/test/progress-bar/ProgressBarSize.test.jsx
@@ -3,19 +3,19 @@ import ProgressBar from '../../../ProgressBar';
 
 const ProgressBarSize = () => (
   <div>
-    <ProgressBar id="progressbarTiny" heightSize="tiny" value={15} color="#8ccc62" />
+    <ProgressBar id="progressbarTiny" heightSize="tiny" valueText="15%" value={15} color="#8ccc62" />
     <br />
     <br />
-    <ProgressBar id="progressbarSmall" heightSize="small" value={30} color="#8ccc62" />
+    <ProgressBar id="progressbarSmall" heightSize="small" valueText="30%" value={30} color="#8ccc62" />
     <br />
     <br />
-    <ProgressBar id="progressbarMedium" heightSize="medium" value={45} color="#8ccc62" />
+    <ProgressBar id="progressbarMedium" heightSize="medium" valueText="45%" value={45} color="#8ccc62" />
     <br />
     <br />
-    <ProgressBar id="progressbarLarge" heightSize="large" value={60} color="#8ccc62" />
+    <ProgressBar id="progressbarLarge" heightSize="large" valueText="60%" value={60} color="#8ccc62" />
     <br />
     <br />
-    <ProgressBar id="progressbarHuge" heightSize="huge" value={75} color="#8ccc62" />
+    <ProgressBar id="progressbarHuge" heightSize="huge" valueText="75%" value={75} color="#8ccc62" />
   </div>
 );