+++ packages/terra-progress-bar/CHANGELOG.md
@@ -3,6 +3,8 @@ Changelog
 
 Unreleased
 ----------
+### Changed
+* Change aria-current value from page to true
 
 2.22.1 - (September 6, 2018)
 ------------------